Changed Power wheel to be handled by middle mouse button, Camera change by Q & E. NPC, Bouncer and Target AI are in Mission_1 level.
There is a bug/error currently where the NPC will gain a null value on Vector location. It's completely random and I'm looking into fixing it. However it seems to be something to do with the blackboard value setting since the object reference is always valid before it continues with the function.